export directly newsletteruser from form to sendingblue
